cleanup chiplist duplication
